+++ packages/sapience/public/sitemap.xml
@@ -1,4 +1,3 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-<sitemap><loc>https://sapience.xyz/sitemap-0.xml</loc></sitemap>
 </sitemapindex>
