+++ packages/sapience/public/sw.js
@@ -1,2 +1,2 @@
-if(!self.define){let e,a={};const i=(i,c)=>(i=new URL(i+".js",c).href,a[i]||new Promise((a=>{if("document"in self){const e=document.createElement("script");e.src=i,e.onload=a,document.head.appendChild(e)}else e=i,importScripts(i),a()})).then((()=>{let e=a[i];if(!e)throw new Error(`Module ${i} didn’t register its module`);return e})));self.define=(c,t)=>{const s=e||("document"in self?document.currentScript.src:"")||location.href;if(a[s])return;let n={};const f=e=>i(e,s),d={module:{uri:s},exports:n,require:f};a[s]=Promise.all(c.map((e=>d[e]||f(e)))).then((e=>(t(...e),n)))}}define(["./workbox-01fd22c6"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/AvenirNextRoundedBold-1080176-export/AvenirNextRoundedBold-1080176.css",revision:"1ceb17790c60a21b9723d3e607aead79"},{url:"/AvenirNextRoundedBold-1080176-export/AvenirNextRoundedBold-1080176.ttf",revision:"ee9e44c0c03af1ffdd399095dbe31442"},{url:"/AvenirNextRoundedBold-1080176-export/AvenirNextRoundedBold-1080176.woff",revision:"7b971a779e02efd771fc017376da708f"},{url:"/AvenirNextRoundedBold-1080176-export/AvenirNextRoundedBold-1080176.woff2",revision:"c8465c4275d797c0949f3b585392eac3"},{url:"/AvenirNextRoundedBoldItalic-1075320-export/AvenirNextRoundedBoldItalic-1075320.css",revision:"1ac4c30afdf37fe83cab27be30cfe663"},{url:"/AvenirNextRoundedBoldItalic-1075320-export/AvenirNextRoundedBoldItalic-1075320.otf",revision:"3f4d1c49e5b4ce229041536366c96764"},{url:"/AvenirNextRoundedBoldItalic-1075320-export/AvenirNextRoundedBoldItalic-1075320.ttf",revision:"f18b943e735bd77c4bf8a6ef47500901"},{url:"/AvenirNextRoundedBoldItalic-1075320-export/AvenirNextRoundedBoldItalic-1075320.woff",revision:"9bb299a1fbb5e76b35ef69b3a18650dd"},{url:"/AvenirNextRoundedBoldItalic-1075320-export/AvenirNextRoundedBoldItalic-1075320.woff2",revision:"43d66773de3fb958bdd054ad6f75eabe"},{url:"/AvenirNextRoundedDemi-1080178-export/AvenirNextRoundedDemi-1080178.css",revision:"43090d76c5fe09130a0bdc44533605ae"},{url:"/AvenirNextRoundedDemi-1080178-export/AvenirNextRoundedDemi-1080178.ttf",revision:"3f13cda3f377e918405094d7d52fc164"},{url:"/AvenirNextRoundedDemi-1080178-export/AvenirNextRoundedDemi-1080178.woff",revision:"34434ce9fcf003c30a65e0f358da9b5a"},{url:"/AvenirNextRoundedDemi-1080178-export/AvenirNextRoundedDemi-1080178.woff2",revision:"cc7595f4dee1699df2168fb8eb497f5b"},{url:"/AvenirNextRoundedDemiItalic-1080179-export/AvenirNextRoundedDemiItalic-1080179.css",revision:"74514acce4b657821cba558fe7d62f09"},{url:"/AvenirNextRoundedDemiItalic-1080179-export/AvenirNextRoundedDemiItalic-1080179.ttf",revision:"4cd4be9e5453e4057286c49e6cdf66d7"},{url:"/AvenirNextRoundedDemiItalic-1080179-export/AvenirNextRoundedDemiItalic-1080179.woff",revision:"4b10054bbe37911428533c11c16bad39"},{url:"/AvenirNextRoundedDemiItalic-1080179-export/AvenirNextRoundedDemiItalic-1080179.woff2",revision:"6fa35f9d865fa0fec67c1d7173af1c38"},{url:"/AvenirNextRoundedItalic-1075329-export/AvenirNextRoundedItalic-1075329.css",revision:"967526b4a331859828c83e0ba1aab1cb"},{url:"/AvenirNextRoundedItalic-1075329-export/AvenirNextRoundedItalic-1075329.otf",revision:"14f13b1eb96472c339066209b25e6e38"},{url:"/AvenirNextRoundedItalic-1075329-export/AvenirNextRoundedItalic-1075329.ttf",revision:"d62a5ea4e6a916c2f85aef1128f5fb6d"},{url:"/AvenirNextRoundedItalic-1075329-export/AvenirNextRoundedItalic-1075329.woff",revision:"a0ae68751e1be333fcbfd84f4a8fa0cd"},{url:"/AvenirNextRoundedItalic-1075329-export/AvenirNextRoundedItalic-1075329.woff2",revision:"cb49a984cc012f60915ecb3d126481f6"},{url:"/AvenirNextRoundedMedium-1080181-export/AvenirNextRoundedMedium-1080181.css",revision:"20de729cfddeafe0f939cb14a474ccbe"},{url:"/AvenirNextRoundedMedium-1080181-export/AvenirNextRoundedMedium-1080181.ttf",revision:"3db056d3eacbed2c61f6ef260893c753"},{url:"/AvenirNextRoundedMedium-1080181-export/AvenirNextRoundedMedium-1080181.woff",revision:"51bb8148be376ffe5ca414724803bd6c"},{url:"/AvenirNextRoundedMedium-1080181-export/AvenirNextRoundedMedium-1080181.woff2",revision:"1cce31aff91395c3d271f16567694d66"},{url:"/AvenirNextRoundedMediumItalic-1080182-export/AvenirNextRoundedMediumItalic-1080182.css",revision:"a141b9630b081a1d819055431aa279ea"},{url:"/AvenirNextRoundedMediumItalic-1080182-export/AvenirNextRoundedMediumItalic-1080182.ttf",revision:"db127d9ee2d611079843192027fdf2ed"},{url:"/AvenirNextRoundedMediumItalic-1080182-export/AvenirNextRoundedMediumItalic-1080182.woff",revision:"41e3a038f73b2165e8999da24b6706b5"},{url:"/AvenirNextRoundedMediumItalic-1080182-export/AvenirNextRoundedMediumItalic-1080182.woff2",revision:"63517191ac08756c53ed1cd8279563c1"},{url:"/AvenirNextRoundedRegular-1080183-export/AvenirNextRoundedRegular-1080183.css",revision:"ad55ff0ae4597806611a805e2474da45"},{url:"/AvenirNextRoundedRegular-1080183-export/AvenirNextRoundedRegular-1080183.ttf",revision:"2e5e8f5eba25e2675b56d4a0ff12b70e"},{url:"/AvenirNextRoundedRegular-1080183-export/AvenirNextRoundedRegular-1080183.woff",revision:"b77d744fb2834b268a252ba377cdc6de"},{url:"/AvenirNextRoundedRegular-1080183-export/AvenirNextRoundedRegular-1080183.woff2",revision:"7e4a9b8881609db0f6a971bcc5a3ef39"},{url:"/_next/app-build-manifest.json",revision:"92e211f5ff4ec6d3105989f1cac8bee3"},{url:"/_next/static/LxaCL6c1p0MwZmi-xfr02/_buildManifest.js",revision:"4e7cba7e679d15c6635270870fb7ba38"},{url:"/_next/static/LxaCL6c1p0MwZmi-xfr02/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/1004.e8196906df9d2d37.js",revision:"e8196906df9d2d37"},{url:"/_next/static/chunks/1004.e8196906df9d2d37.js.map",revision:"22875d43a9eac4e0b326092b701ed1fd"},{url:"/_next/static/chunks/1313.d35a3ff320d0590c.js",revision:"d35a3ff320d0590c"},{url:"/_next/static/chunks/1313.d35a3ff320d0590c.js.map",revision:"e16d297a6012599f412472667597d855"},{url:"/_next/static/chunks/1379.237c4b79d8d7b465.js",revision:"237c4b79d8d7b465"},{url:"/_next/static/chunks/1379.237c4b79d8d7b465.js.map",revision:"dd3f687de2240e7a27599e75faa962be"},{url:"/_next/static/chunks/1387-8c6452f3f375ca5c.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/1387-8c6452f3f375ca5c.js.map",revision:"34b8205b888274a204607f76b3367e69"},{url:"/_next/static/chunks/1431.3c48a3c16aab9f28.js",revision:"3c48a3c16aab9f28"},{url:"/_next/static/chunks/1431.3c48a3c16aab9f28.js.map",revision:"d3523ab8238be30098c126075a361af5"},{url:"/_next/static/chunks/1491-54e76d30e871efdd.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/1491-54e76d30e871efdd.js.map",revision:"55a6566aca1d32e7a56dfc911c159e6b"},{url:"/_next/static/chunks/1612.8cb7f823c221fd24.js",revision:"8cb7f823c221fd24"},{url:"/_next/static/chunks/1612.8cb7f823c221fd24.js.map",revision:"144fea08abfc3ad454d1d092a13edc31"},{url:"/_next/static/chunks/183.40f993f72fbe8d7e.js",revision:"40f993f72fbe8d7e"},{url:"/_next/static/chunks/183.40f993f72fbe8d7e.js.map",revision:"47e48c5d9040fafa4b85f65edc19bfa7"},{url:"/_next/static/chunks/1843.1225c1e9862eabd6.js",revision:"1225c1e9862eabd6"},{url:"/_next/static/chunks/1843.1225c1e9862eabd6.js.map",revision:"aa95849264dc3b7f97e225fbba20eef3"},{url:"/_next/static/chunks/18b16e15-c7fa9a10f549bc53.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/18b16e15-c7fa9a10f549bc53.js.map",revision:"4edb8068b4acecfa5edd70cb6f06a2bf"},{url:"/_next/static/chunks/2058.443982227ed90ff2.js",revision:"443982227ed90ff2"},{url:"/_next/static/chunks/2058.443982227ed90ff2.js.map",revision:"406bc63add6341c45ffee4e099653a75"},{url:"/_next/static/chunks/2084.721969af66f6a1b4.js",revision:"721969af66f6a1b4"},{url:"/_next/static/chunks/2084.721969af66f6a1b4.js.map",revision:"7074c54c1683cafca1f43c7b173dca7b"},{url:"/_next/static/chunks/2501-428399479317876d.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/2501-428399479317876d.js.map",revision:"c20068574ed3d1b7861bb01af7890086"},{url:"/_next/static/chunks/2513.f4261cbc4c5ca2e7.js",revision:"f4261cbc4c5ca2e7"},{url:"/_next/static/chunks/2513.f4261cbc4c5ca2e7.js.map",revision:"a68fad71facfd946322c9f3118294463"},{url:"/_next/static/chunks/2527.3d63cba5a47c5542.js",revision:"3d63cba5a47c5542"},{url:"/_next/static/chunks/2527.3d63cba5a47c5542.js.map",revision:"862e8457f9bd000c7b9d7ea60ab74d42"},{url:"/_next/static/chunks/3009.28d3a81601309cf8.js",revision:"28d3a81601309cf8"},{url:"/_next/static/chunks/3009.28d3a81601309cf8.js.map",revision:"f2a7cfcf2eb2ad9c3a4a8ebd0f25ae5d"},{url:"/_next/static/chunks/3106.3428e2aef5aae2fe.js",revision:"3428e2aef5aae2fe"},{url:"/_next/static/chunks/3106.3428e2aef5aae2fe.js.map",revision:"07152bd06cf2f705f4cb4ee677f3bc1a"},{url:"/_next/static/chunks/3178.14ed89786c7fdaf1.js",revision:"14ed89786c7fdaf1"},{url:"/_next/static/chunks/3178.14ed89786c7fdaf1.js.map",revision:"107a36cee548db39c7887e490e4adb13"},{url:"/_next/static/chunks/3470.738ba65f2e46a7e5.js",revision:"738ba65f2e46a7e5"},{url:"/_next/static/chunks/3470.738ba65f2e46a7e5.js.map",revision:"53877601621d7dc5e0e4e30313b8e118"},{url:"/_next/static/chunks/3485.fec580dec970a563.js",revision:"fec580dec970a563"},{url:"/_next/static/chunks/3485.fec580dec970a563.js.map",revision:"4c7ef04bdcc69881633dd45e3bec91da"},{url:"/_next/static/chunks/35.cc63778b09482458.js",revision:"cc63778b09482458"},{url:"/_next/static/chunks/35.cc63778b09482458.js.map",revision:"cf2846581eee5a591d8a4a592532dd9a"},{url:"/_next/static/chunks/3527-92c4af87d7544430.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/3527-92c4af87d7544430.js.map",revision:"ba4cf6ad5a625b3e247bb22af89f75af"},{url:"/_next/static/chunks/3761.f390d71a8c3a3f30.js",revision:"f390d71a8c3a3f30"},{url:"/_next/static/chunks/3761.f390d71a8c3a3f30.js.map",revision:"93b37ac1053354802ff83550e1decb96"},{url:"/_next/static/chunks/3813.265bbdba1e5e3e50.js",revision:"265bbdba1e5e3e50"},{url:"/_next/static/chunks/3813.265bbdba1e5e3e50.js.map",revision:"e9e900d3d1f52e0895ff10318d7778cd"},{url:"/_next/static/chunks/3840.3a74d7a9662cf0f2.js",revision:"3a74d7a9662cf0f2"},{url:"/_next/static/chunks/3840.3a74d7a9662cf0f2.js.map",revision:"4222e68e33f85e8ce65597c156f23ceb"},{url:"/_next/static/chunks/3930.84d81ab69f8e184e.js",revision:"84d81ab69f8e184e"},{url:"/_next/static/chunks/3930.84d81ab69f8e184e.js.map",revision:"d41abe9eda22a9ebd03d9dd0fe5f857c"},{url:"/_next/static/chunks/4129.e10e5225439a8302.js",revision:"e10e5225439a8302"},{url:"/_next/static/chunks/4129.e10e5225439a8302.js.map",revision:"5eaca6406bf71103b870a9e6d4c614b2"},{url:"/_next/static/chunks/4311.9928c50563c60a8d.js",revision:"9928c50563c60a8d"},{url:"/_next/static/chunks/4311.9928c50563c60a8d.js.map",revision:"007cbb36ac60c6074c8c8aef20e084f5"},{url:"/_next/static/chunks/432.307faf39b8d17ac8.js",revision:"307faf39b8d17ac8"},{url:"/_next/static/chunks/432.307faf39b8d17ac8.js.map",revision:"32f15981a534ba4f3adb79b29d146cfe"},{url:"/_next/static/chunks/4390.456f64929fd37c44.js",revision:"456f64929fd37c44"},{url:"/_next/static/chunks/4390.456f64929fd37c44.js.map",revision:"5cb5eb19863ef582838c20cca419bfb9"},{url:"/_next/static/chunks/474.cf75ea0ef0d83bdc.js",revision:"cf75ea0ef0d83bdc"},{url:"/_next/static/chunks/474.cf75ea0ef0d83bdc.js.map",revision:"f450a6c433a394ac6d266353ef205f45"},{url:"/_next/static/chunks/4928.e06783c962e02092.js",revision:"e06783c962e02092"},{url:"/_next/static/chunks/4928.e06783c962e02092.js.map",revision:"997500d984bb206015942768dddb0af7"},{url:"/_next/static/chunks/5061.94bb1981ba994dda.js",revision:"94bb1981ba994dda"},{url:"/_next/static/chunks/5061.94bb1981ba994dda.js.map",revision:"e2f0019a3bdb7f91431333c8087ac116"},{url:"/_next/static/chunks/5154.0acb602e397df24b.js",revision:"0acb602e397df24b"},{url:"/_next/static/chunks/5154.0acb602e397df24b.js.map",revision:"f7fa2e58c5648b79ce6e33b0434afe55"},{url:"/_next/static/chunks/5262.ab52f5bb275e3203.js",revision:"ab52f5bb275e3203"},{url:"/_next/static/chunks/5262.ab52f5bb275e3203.js.map",revision:"57a654cb3bb40ddf4f97f4f59dd68e32"},{url:"/_next/static/chunks/5594-64ccb7bb8dd7fa23.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/5594-64ccb7bb8dd7fa23.js.map",revision:"4413fe8191c305c27b6647bc9c0c9add"},{url:"/_next/static/chunks/5698.d407a957bdda6088.js",revision:"d407a957bdda6088"},{url:"/_next/static/chunks/5698.d407a957bdda6088.js.map",revision:"ab9b7d64851a2511a2c2b0c28ec17846"},{url:"/_next/static/chunks/5761.045122dd9e68eeda.js",revision:"045122dd9e68eeda"},{url:"/_next/static/chunks/5761.045122dd9e68eeda.js.map",revision:"ef1b6fdf13de7ee2a16c24099cd666a4"},{url:"/_next/static/chunks/5842.646ce968e2d41fa4.js",revision:"646ce968e2d41fa4"},{url:"/_next/static/chunks/5988.b0efe90b26384a5c.js",revision:"b0efe90b26384a5c"},{url:"/_next/static/chunks/5988.b0efe90b26384a5c.js.map",revision:"acf68c9785b9e0717f81c3def1c71b15"},{url:"/_next/static/chunks/6066.c31de60e4fe6f81f.js",revision:"c31de60e4fe6f81f"},{url:"/_next/static/chunks/6066.c31de60e4fe6f81f.js.map",revision:"7fc69b8d07d73b0e32df3b59bd277aee"},{url:"/_next/static/chunks/6221-59f57820f302b796.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/6221-59f57820f302b796.js.map",revision:"aac4fb2fe02a76cd5d6ab1b0e002c962"},{url:"/_next/static/chunks/6322-8d752e1365d19225.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/6322-8d752e1365d19225.js.map",revision:"eaef247adf910776df168aade78aeebf"},{url:"/_next/static/chunks/645.db654d3ed69e2e63.js",revision:"db654d3ed69e2e63"},{url:"/_next/static/chunks/645.db654d3ed69e2e63.js.map",revision:"185c719a039ef2c5c4712c04a034373d"},{url:"/_next/static/chunks/6482.5ab67833f8c83b1d.js",revision:"5ab67833f8c83b1d"},{url:"/_next/static/chunks/6482.5ab67833f8c83b1d.js.map",revision:"e4824573c76dc09ac8058a124f7205f7"},{url:"/_next/static/chunks/6527.c8e348452104a7cf.js",revision:"c8e348452104a7cf"},{url:"/_next/static/chunks/6527.c8e348452104a7cf.js.map",revision:"78f01862509a6e11f11d0e7df2002c38"},{url:"/_next/static/chunks/6892.4411a0bc75f0cfe1.js",revision:"4411a0bc75f0cfe1"},{url:"/_next/static/chunks/6892.4411a0bc75f0cfe1.js.map",revision:"02e3334c915b52ec54621ea57baf2847"},{url:"/_next/static/chunks/6923.3d83b76f2a8d0253.js",revision:"3d83b76f2a8d0253"},{url:"/_next/static/chunks/6923.3d83b76f2a8d0253.js.map",revision:"446fa804e43d45f2f5257c8f475ab074"},{url:"/_next/static/chunks/7074.35929822a1976bef.js",revision:"35929822a1976bef"},{url:"/_next/static/chunks/7074.35929822a1976bef.js.map",revision:"22b38433242af691d811751c0cda1751"},{url:"/_next/static/chunks/7569.b7bd65468d3829ed.js",revision:"b7bd65468d3829ed"},{url:"/_next/static/chunks/7569.b7bd65468d3829ed.js.map",revision:"a698e8ed75d21018a03b868ce1da7803"},{url:"/_next/static/chunks/7583.7797feaa833f5d23.js",revision:"7797feaa833f5d23"},{url:"/_next/static/chunks/7583.7797feaa833f5d23.js.map",revision:"9d90b8b82140f5bfc4f5b247233a9c74"},{url:"/_next/static/chunks/7617.b8d2feff1c6d13a6.js",revision:"b8d2feff1c6d13a6"},{url:"/_next/static/chunks/7617.b8d2feff1c6d13a6.js.map",revision:"6bae3b355b34f0aa0354e6b64588ebf4"},{url:"/_next/static/chunks/7788.3ad7ad9f1e7f7833.js",revision:"3ad7ad9f1e7f7833"},{url:"/_next/static/chunks/7788.3ad7ad9f1e7f7833.js.map",revision:"cf6c31d0b4727880dbb21e36c2b2cb5b"},{url:"/_next/static/chunks/7874.89fb9eeb6be31224.js",revision:"89fb9eeb6be31224"},{url:"/_next/static/chunks/7874.89fb9eeb6be31224.js.map",revision:"20f50cd0532442d9984ccf1a230bd356"},{url:"/_next/static/chunks/790.13f6c4b9ff1791fd.js",revision:"13f6c4b9ff1791fd"},{url:"/_next/static/chunks/790.13f6c4b9ff1791fd.js.map",revision:"43b0384d0af5578b72bf2353ce8248e5"},{url:"/_next/static/chunks/7945.8067e2fa1c882f97.js",revision:"8067e2fa1c882f97"},{url:"/_next/static/chunks/7945.8067e2fa1c882f97.js.map",revision:"25e7f04d95da029e45237921184045eb"},{url:"/_next/static/chunks/7963-917c7bee629b26c4.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/7971-5d26e6a66975079f.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/7971-5d26e6a66975079f.js.map",revision:"78bebd8c6f94a0ab689a2b60f821ea99"},{url:"/_next/static/chunks/7e28f48e-5d0313d8310c9156.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/7e28f48e-5d0313d8310c9156.js.map",revision:"d144d0618337dd9843b905d705260304"},{url:"/_next/static/chunks/8057.aa273303cb883692.js",revision:"aa273303cb883692"},{url:"/_next/static/chunks/8057.aa273303cb883692.js.map",revision:"a0e76f12b7f9009a66ea3e8d6f56b584"},{url:"/_next/static/chunks/8250.83b4198009f06312.js",revision:"83b4198009f06312"},{url:"/_next/static/chunks/8250.83b4198009f06312.js.map",revision:"afb6d5be266b24f2f078b246bdee6a93"},{url:"/_next/static/chunks/830.bae1ca90def593f2.js",revision:"bae1ca90def593f2"},{url:"/_next/static/chunks/830.bae1ca90def593f2.js.map",revision:"105063600626d28f39239bd788a10f4d"},{url:"/_next/static/chunks/8456.25e5d2cb26ee031a.js",revision:"25e5d2cb26ee031a"},{url:"/_next/static/chunks/8456.25e5d2cb26ee031a.js.map",revision:"3296afb797c8d246ab3c7291ce345a67"},{url:"/_next/static/chunks/8517.5d60ccbb8c18c647.js",revision:"5d60ccbb8c18c647"},{url:"/_next/static/chunks/8517.5d60ccbb8c18c647.js.map",revision:"8096cfaed153334962c681831d0ea56a"},{url:"/_next/static/chunks/8624.69bca3b30ceeab59.js",revision:"69bca3b30ceeab59"},{url:"/_next/static/chunks/8624.69bca3b30ceeab59.js.map",revision:"8710a1c28bd36689244545e69e87ecd4"},{url:"/_next/static/chunks/8750.be187129cb2ee251.js",revision:"be187129cb2ee251"},{url:"/_next/static/chunks/8750.be187129cb2ee251.js.map",revision:"90f65370dba730445d90afcd04265d60"},{url:"/_next/static/chunks/8791.1fa8b4b35170ab7c.js",revision:"1fa8b4b35170ab7c"},{url:"/_next/static/chunks/8791.1fa8b4b35170ab7c.js.map",revision:"f378c498f166e67e104e9ad05eb373fe"},{url:"/_next/static/chunks/8811-28713fb5fc5f5fcf.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/8811-28713fb5fc5f5fcf.js.map",revision:"e5d737e0d77535a3f973e2e9d5a0fa4c"},{url:"/_next/static/chunks/8997.9487f2b0653c146f.js",revision:"9487f2b0653c146f"},{url:"/_next/static/chunks/8997.9487f2b0653c146f.js.map",revision:"9d71c4c256f17077d12824c072afd9ae"},{url:"/_next/static/chunks/9058.681a93fdcfdf8ef5.js",revision:"681a93fdcfdf8ef5"},{url:"/_next/static/chunks/9058.681a93fdcfdf8ef5.js.map",revision:"4d89158fa99909704b5c68a8af7ba690"},{url:"/_next/static/chunks/9154-790f23abccdb3e49.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/9154-790f23abccdb3e49.js.map",revision:"8e2d49c5257d72b24ec4f6ed3e9e93fb"},{url:"/_next/static/chunks/9536.220bd442696c7036.js",revision:"220bd442696c7036"},{url:"/_next/static/chunks/9536.220bd442696c7036.js.map",revision:"0dfefbae126fa34db207d499ad650179"},{url:"/_next/static/chunks/9678-778ff3b23c04f64f.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/9678-778ff3b23c04f64f.js.map",revision:"7ae7be6082a2c3dbb7da8991d4c5aa1e"},{url:"/_next/static/chunks/9989.a8b158d927b15331.js",revision:"a8b158d927b15331"},{url:"/_next/static/chunks/9989.a8b158d927b15331.js.map",revision:"86a7b22cdf1a6922310a7a739e986cfc"},{url:"/_next/static/chunks/a31ab248-eed7c6516d748d54.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/a31ab248-eed7c6516d748d54.js.map",revision:"af2c5539be46e9b7ce28f3efaa83bfd1"},{url:"/_next/static/chunks/app/_not-found/page-f27100c7951df8d1.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/app/bots/page-c4088696a7be0507.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/app/bots/page-c4088696a7be0507.js.map",revision:"a34503e9fd91a3557bbea70d431a2ae9"},{url:"/_next/static/chunks/app/error-1ebdad38a45beb59.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/app/error-1ebdad38a45beb59.js.map",revision:"85276708ce8e8b27ee9d264c6905b9ca"},{url:"/_next/static/chunks/app/forecasting/%5BchainId%5D/%5BmarketAddress%5D/page-d1e2ec050a2ce001.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/app/forecasting/%5BchainId%5D/%5BmarketAddress%5D/page-d1e2ec050a2ce001.js.map",revision:"abf930c38e0fe029d99229498fabfc7c"},{url:"/_next/static/chunks/app/forecasting/page-20bbacd929f458cc.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/app/forecasting/page-20bbacd929f458cc.js.map",revision:"0a3d2e9f7a37a8c8a494109876d8b9ef"},{url:"/_next/static/chunks/app/futarchy/page-a77f931d9c55494c.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/app/futarchy/page-a77f931d9c55494c.js.map",revision:"c75ceb3e4187cee449a6cdbe845fb705"},{url:"/_next/static/chunks/app/global-error-918feb010657bae6.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/app/global-error-918feb010657bae6.js.map",revision:"c93f71e90d79a6aa53384d08ac9d9bb7"},{url:"/_next/static/chunks/app/layout-225697953947ea34.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/app/layout-225697953947ea34.js.map",revision:"eeb9c82d87dfa56b659f60585cacc57b"},{url:"/_next/static/chunks/app/leaderboard/%5Bid%5D/epochs/%5Bepoch%5D/page-f165f2c8250995b5.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/app/leaderboard/%5Bid%5D/epochs/%5Bepoch%5D/page-f165f2c8250995b5.js.map",revision:"321acd41b0c21df45ced6c8a9c40a61e"},{url:"/_next/static/chunks/app/leaderboard/page-bacbd1ac2b31c3ed.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/app/leaderboard/page-bacbd1ac2b31c3ed.js.map",revision:"5c826c17db9b1ad237d3f3d01fe3bf3c"},{url:"/_next/static/chunks/app/not-found-d03852b57c10b5ea.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/app/page-ddc6eb1ceaf37d6f.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/app/page-ddc6eb1ceaf37d6f.js.map",revision:"3fec612a6727488c25660072bb963144"},{url:"/_next/static/chunks/f4dc6ce0-7d93a3b742b388d5.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/f4dc6ce0-7d93a3b742b388d5.js.map",revision:"b41ec79bb0c7bec23559ecbd03b2c9a7"},{url:"/_next/static/chunks/framework-1625a25bd56a487b.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/framework-1625a25bd56a487b.js.map",revision:"b61dec9ba5e7fc11d2aac637da9a1856"},{url:"/_next/static/chunks/main-app-ac55f851f27ea83e.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/main-app-ac55f851f27ea83e.js.map",revision:"10e6d1b96e6db4b7c7d266cf4d5b3aa9"},{url:"/_next/static/chunks/main-b66bab6e4ca24744.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/main-b66bab6e4ca24744.js.map",revision:"2baed322d168e42b8b8ebafa1bdf3820"},{url:"/_next/static/chunks/pages/_app-101424800861aca5.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/pages/_app-101424800861aca5.js.map",revision:"04fb8e6330016ae0fc3020cdf57df2c8"},{url:"/_next/static/chunks/pages/_error-456bc4d02b78646c.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/pages/_error-456bc4d02b78646c.js.map",revision:"ad96e20c79d82bfb66cc1842e813cde9"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-976c5fbb14db1808.js",revision:"LxaCL6c1p0MwZmi-xfr02"},{url:"/_next/static/chunks/webpack-976c5fbb14db1808.js.map",revision:"8cb1b74930d36122c41a80177bad3a63"},{url:"/_next/static/css/01e303d912fdd7db.css",revision:"01e303d912fdd7db"},{url:"/_next/static/css/01e303d912fdd7db.css.map",revision:"e2968a09346b76b0991e5b24bfe65214"},{url:"/_next/static/css/59d66467b5973e01.css",revision:"59d66467b5973e01"},{url:"/_next/static/css/59d66467b5973e01.css.map",revision:"0c70a9a37f3cec399c559fa383dfbb5a"},{url:"/_next/static/css/8073d4869bb3d095.css",revision:"8073d4869bb3d095"},{url:"/_next/static/css/8073d4869bb3d095.css.map",revision:"f7936320d2959b9143c53bf2cd67acad"},{url:"/_next/static/media/font.0d99608a.woff",revision:"0d99608a"},{url:"/_next/static/media/font.1437a347.woff2",revision:"1437a347"},{url:"/_next/static/media/font.15f0cb14.woff2",revision:"15f0cb14"},{url:"/_next/static/media/font.172826d2.woff2",revision:"172826d2"},{url:"/_next/static/media/font.21d6c94e.woff2",revision:"21d6c94e"},{url:"/_next/static/media/font.2484265e.woff",revision:"2484265e"},{url:"/_next/static/media/font.248fd520.woff",revision:"248fd520"},{url:"/_next/static/media/font.2913dd65.woff",revision:"2913dd65"},{url:"/_next/static/media/font.29904e4a.woff2",revision:"29904e4a"},{url:"/_next/static/media/font.2c3dda54.woff",revision:"2c3dda54"},{url:"/_next/static/media/font.2d75bfb1.woff2",revision:"2d75bfb1"},{url:"/_next/static/media/font.2f474a58.woff2",revision:"2f474a58"},{url:"/_next/static/media/font.2f548f0d.woff2",revision:"2f548f0d"},{url:"/_next/static/media/font.2fa17060.woff2",revision:"2fa17060"},{url:"/_next/static/media/font.3057c2b7.woff",revision:"3057c2b7"},{url:"/_next/static/media/font.3108bbad.woff",revision:"3108bbad"},{url:"/_next/static/media/font.31ed5aac.woff2",revision:"31ed5aac"},{url:"/_next/static/media/font.37cdad6c.woff2",revision:"37cdad6c"},{url:"/_next/static/media/font.39d2f5c6.woff2",revision:"39d2f5c6"},{url:"/_next/static/media/font.435bd86c.woff2",revision:"435bd86c"},{url:"/_next/static/media/font.4b393fab.woff",revision:"4b393fab"},{url:"/_next/static/media/font.50bbef7d.woff",revision:"50bbef7d"},{url:"/_next/static/media/font.553b8400.woff",revision:"553b8400"},{url:"/_next/static/media/font.5cf8a729.woff2",revision:"5cf8a729"},{url:"/_next/static/media/font.5d6a72af.woff2",revision:"5d6a72af"},{url:"/_next/static/media/font.635bb62c.woff",revision:"635bb62c"},{url:"/_next/static/media/font.6509c247.woff",revision:"6509c247"},{url:"/_next/static/media/font.6cd92d24.woff2",revision:"6cd92d24"},{url:"/_next/static/media/font.73a9d55f.woff",revision:"73a9d55f"},{url:"/_next/static/media/font.7bcc2af9.woff",revision:"7bcc2af9"},{url:"/_next/static/media/font.7c4c2bf8.woff2",revision:"7c4c2bf8"},{url:"/_next/static/media/font.7c7cecdb.woff2",revision:"7c7cecdb"},{url:"/_next/static/media/font.816c5319.woff",revision:"816c5319"},{url:"/_next/static/media/font.82bbad3b.woff2",revision:"82bbad3b"},{url:"/_next/static/media/font.856572dd.woff2",revision:"856572dd"},{url:"/_next/static/media/font.8bb377a0.woff",revision:"8bb377a0"},{url:"/_next/static/media/font.8ee62147.woff2",revision:"8ee62147"},{url:"/_next/static/media/font.8f4ab0ae.woff2",revision:"8f4ab0ae"},{url:"/_next/static/media/font.9491bc5f.woff2",revision:"9491bc5f"},{url:"/_next/static/media/font.9c12eced.woff",revision:"9c12eced"},{url:"/_next/static/media/font.9d17a49d.woff",revision:"9d17a49d"},{url:"/_next/static/media/font.9fb5eaf7.woff",revision:"9fb5eaf7"},{url:"/_next/static/media/font.9fd6a2c2.woff",revision:"9fd6a2c2"},{url:"/_next/static/media/font.a11bbcef.woff",revision:"a11bbcef"},{url:"/_next/static/media/font.a3e54e3d.woff",revision:"a3e54e3d"},{url:"/_next/static/media/font.a4380fcb.woff",revision:"a4380fcb"},{url:"/_next/static/media/font.a7183547.woff2",revision:"a7183547"},{url:"/_next/static/media/font.a7d755eb.woff",revision:"a7d755eb"},{url:"/_next/static/media/font.afa2688e.woff",revision:"afa2688e"},{url:"/_next/static/media/font.b3f27314.woff",revision:"b3f27314"},{url:"/_next/static/media/font.b850aef4.woff2",revision:"b850aef4"},{url:"/_next/static/media/font.b869ae38.woff",revision:"b869ae38"},{url:"/_next/static/media/font.b99947cf.woff",revision:"b99947cf"},{url:"/_next/static/media/font.ba404641.woff2",revision:"ba404641"},{url:"/_next/static/media/font.bdffa644.woff2",revision:"bdffa644"},{url:"/_next/static/media/font.be02cff3.woff",revision:"be02cff3"},{url:"/_next/static/media/font.bf00c3fd.woff2",revision:"bf00c3fd"},{url:"/_next/static/media/font.c0a46350.woff2",revision:"c0a46350"},{url:"/_next/static/media/font.c9641110.woff",revision:"c9641110"},{url:"/_next/static/media/font.cb0db050.woff2",revision:"cb0db050"},{url:"/_next/static/media/font.d0a69e17.woff2",revision:"d0a69e17"},{url:"/_next/static/media/font.d18ef773.woff",revision:"d18ef773"},{url:"/_next/static/media/font.d3118412.woff2",revision:"d3118412"},{url:"/_next/static/media/font.d63f196b.woff2",revision:"d63f196b"},{url:"/_next/static/media/font.d8425c1f.woff2",revision:"d8425c1f"},{url:"/_next/static/media/font.d890d2d6.woff",revision:"d890d2d6"},{url:"/_next/static/media/font.e2b7d353.woff2",revision:"e2b7d353"},{url:"/_next/static/media/font.e6399456.woff2",revision:"e6399456"},{url:"/_next/static/media/font.e7691173.woff2",revision:"e7691173"},{url:"/_next/static/media/font.e93302c2.woff",revision:"e93302c2"},{url:"/_next/static/media/font.e985e56f.woff",revision:"e985e56f"},{url:"/_next/static/media/font.e9e26597.woff2",revision:"e9e26597"},{url:"/_next/static/media/font.f0702b0f.woff",revision:"f0702b0f"},{url:"/_next/static/media/font.f1b46486.woff2",revision:"f1b46486"},{url:"/_next/static/media/font.f2fca399.woff",revision:"f2fca399"},{url:"/_next/static/media/font.f6457293.woff",revision:"f6457293"},{url:"/_next/static/media/font.f8d712aa.woff",revision:"f8d712aa"},{url:"/_next/static/media/font.faad02c9.woff",revision:"faad02c9"},{url:"/_next/static/media/font.fd1ecef3.woff",revision:"fd1ecef3"},{url:"/_next/static/media/font.fffd047e.woff2",revision:"fffd047e"},{url:"/claude.mp4",revision:"c8753825822cb0d863e9cb356b828217"},{url:"/etherscan.svg",revision:"7b7de5f72960e6da7ea41f35c3bd19da"},{url:"/favicon.svg",revision:"6f9c4e4689886cdc55e2ba1c52ce6cad"},{url:"/footer.png",revision:"7e3d75771110e9747f3a36dcbe4e9d2c"},{url:"/futarchy.png",revision:"d9695200cbd9d9d3eb25fbb8968b5b27"},{url:"/icons/icon-128x128.png",revision:"47b15e1ccdc4cab159cd43bfd1b8a208"},{url:"/icons/icon-144x144.png",revision:"91bd0b770d8c08e0517108dd20b5e09a"},{url:"/icons/icon-152x152.png",revision:"92e01007277fe4078fc585ad021f121f"},{url:"/icons/icon-192x192.png",revision:"007db62f745fd4b1cb63a1bac81d7b66"},{url:"/icons/icon-384x384.png",revision:"e481950d9351fa3b5149db414dff4144"},{url:"/icons/icon-512x512.png",revision:"97e635e3cc9a470518cd09ae80b2045d"},{url:"/icons/icon-72x72.png",revision:"3917754c4d089732711aa1324af449ff"},{url:"/icons/icon-96x96.png",revision:"44c96882dfc1f28a48b7fab96ebde9a6"},{url:"/logo.svg",revision:"fa07e01a214f11b798a83d5e5d79edc3"},{url:"/logomark.svg",revision:"bd5157f0874637ad15dae13d8893c8e4"},{url:"/lottie/logomark.json",revision:"42a379ce014eeae8baf3e838e3e953ac"},{url:"/manifest.json",revision:"a7d93b6cb0ee35bcc458602e953fda9f"},{url:"/og-image.png",revision:"a35a45c8a2a35e3df1fb612ea7a47e50"},{url:"/robots.txt",revision:"375f55efde6139766b4966f3299897d0"},{url:"/sitemap-0.xml",revision:"e4ef3c1ab973c940725d938b2b8ea3b9"},{url:"/sitemap.xml",revision:"115de904005dcab1401b4ed9223ab2f8"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:a,event:i,state:c})=>a&&"opaqueredirect"===a.type?new Response(a.body,{status:200,statusText:"OK",headers:a.headers}):a}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const a=e.pathname;return!a.startsWith("/api/auth/")&&!!a.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
+if(!self.define){let e,a={};const i=(i,c)=>(i=new URL(i+".js",c).href,a[i]||new Promise((a=>{if("document"in self){const e=document.createElement("script");e.src=i,e.onload=a,document.head.appendChild(e)}else e=i,importScripts(i),a()})).then((()=>{let e=a[i];if(!e)throw new Error(`Module ${i} didn’t register its module`);return e})));self.define=(c,t)=>{const s=e||("document"in self?document.currentScript.src:"")||location.href;if(a[s])return;let n={};const f=e=>i(e,s),d={module:{uri:s},exports:n,require:f};a[s]=Promise.all(c.map((e=>d[e]||f(e)))).then((e=>(t(...e),n)))}}define(["./workbox-01fd22c6"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/AvenirNextRoundedBold-1080176-export/AvenirNextRoundedBold-1080176.css",revision:"1ceb17790c60a21b9723d3e607aead79"},{url:"/AvenirNextRoundedBold-1080176-export/AvenirNextRoundedBold-1080176.ttf",revision:"ee9e44c0c03af1ffdd399095dbe31442"},{url:"/AvenirNextRoundedBold-1080176-export/AvenirNextRoundedBold-1080176.woff",revision:"7b971a779e02efd771fc017376da708f"},{url:"/AvenirNextRoundedBold-1080176-export/AvenirNextRoundedBold-1080176.woff2",revision:"c8465c4275d797c0949f3b585392eac3"},{url:"/AvenirNextRoundedBoldItalic-1075320-export/AvenirNextRoundedBoldItalic-1075320.css",revision:"1ac4c30afdf37fe83cab27be30cfe663"},{url:"/AvenirNextRoundedBoldItalic-1075320-export/AvenirNextRoundedBoldItalic-1075320.otf",revision:"3f4d1c49e5b4ce229041536366c96764"},{url:"/AvenirNextRoundedBoldItalic-1075320-export/AvenirNextRoundedBoldItalic-1075320.ttf",revision:"f18b943e735bd77c4bf8a6ef47500901"},{url:"/AvenirNextRoundedBoldItalic-1075320-export/AvenirNextRoundedBoldItalic-1075320.woff",revision:"9bb299a1fbb5e76b35ef69b3a18650dd"},{url:"/AvenirNextRoundedBoldItalic-1075320-export/AvenirNextRoundedBoldItalic-1075320.woff2",revision:"43d66773de3fb958bdd054ad6f75eabe"},{url:"/AvenirNextRoundedDemi-1080178-export/AvenirNextRoundedDemi-1080178.css",revision:"43090d76c5fe09130a0bdc44533605ae"},{url:"/AvenirNextRoundedDemi-1080178-export/AvenirNextRoundedDemi-1080178.ttf",revision:"3f13cda3f377e918405094d7d52fc164"},{url:"/AvenirNextRoundedDemi-1080178-export/AvenirNextRoundedDemi-1080178.woff",revision:"34434ce9fcf003c30a65e0f358da9b5a"},{url:"/AvenirNextRoundedDemi-1080178-export/AvenirNextRoundedDemi-1080178.woff2",revision:"cc7595f4dee1699df2168fb8eb497f5b"},{url:"/AvenirNextRoundedDemiItalic-1080179-export/AvenirNextRoundedDemiItalic-1080179.css",revision:"74514acce4b657821cba558fe7d62f09"},{url:"/AvenirNextRoundedDemiItalic-1080179-export/AvenirNextRoundedDemiItalic-1080179.ttf",revision:"4cd4be9e5453e4057286c49e6cdf66d7"},{url:"/AvenirNextRoundedDemiItalic-1080179-export/AvenirNextRoundedDemiItalic-1080179.woff",revision:"4b10054bbe37911428533c11c16bad39"},{url:"/AvenirNextRoundedDemiItalic-1080179-export/AvenirNextRoundedDemiItalic-1080179.woff2",revision:"6fa35f9d865fa0fec67c1d7173af1c38"},{url:"/AvenirNextRoundedItalic-1075329-export/AvenirNextRoundedItalic-1075329.css",revision:"967526b4a331859828c83e0ba1aab1cb"},{url:"/AvenirNextRoundedItalic-1075329-export/AvenirNextRoundedItalic-1075329.otf",revision:"14f13b1eb96472c339066209b25e6e38"},{url:"/AvenirNextRoundedItalic-1075329-export/AvenirNextRoundedItalic-1075329.ttf",revision:"d62a5ea4e6a916c2f85aef1128f5fb6d"},{url:"/AvenirNextRoundedItalic-1075329-export/AvenirNextRoundedItalic-1075329.woff",revision:"a0ae68751e1be333fcbfd84f4a8fa0cd"},{url:"/AvenirNextRoundedItalic-1075329-export/AvenirNextRoundedItalic-1075329.woff2",revision:"cb49a984cc012f60915ecb3d126481f6"},{url:"/AvenirNextRoundedMedium-1080181-export/AvenirNextRoundedMedium-1080181.css",revision:"20de729cfddeafe0f939cb14a474ccbe"},{url:"/AvenirNextRoundedMedium-1080181-export/AvenirNextRoundedMedium-1080181.ttf",revision:"3db056d3eacbed2c61f6ef260893c753"},{url:"/AvenirNextRoundedMedium-1080181-export/AvenirNextRoundedMedium-1080181.woff",revision:"51bb8148be376ffe5ca414724803bd6c"},{url:"/AvenirNextRoundedMedium-1080181-export/AvenirNextRoundedMedium-1080181.woff2",revision:"1cce31aff91395c3d271f16567694d66"},{url:"/AvenirNextRoundedMediumItalic-1080182-export/AvenirNextRoundedMediumItalic-1080182.css",revision:"a141b9630b081a1d819055431aa279ea"},{url:"/AvenirNextRoundedMediumItalic-1080182-export/AvenirNextRoundedMediumItalic-1080182.ttf",revision:"db127d9ee2d611079843192027fdf2ed"},{url:"/AvenirNextRoundedMediumItalic-1080182-export/AvenirNextRoundedMediumItalic-1080182.woff",revision:"41e3a038f73b2165e8999da24b6706b5"},{url:"/AvenirNextRoundedMediumItalic-1080182-export/AvenirNextRoundedMediumItalic-1080182.woff2",revision:"63517191ac08756c53ed1cd8279563c1"},{url:"/AvenirNextRoundedRegular-1080183-export/AvenirNextRoundedRegular-1080183.css",revision:"ad55ff0ae4597806611a805e2474da45"},{url:"/AvenirNextRoundedRegular-1080183-export/AvenirNextRoundedRegular-1080183.ttf",revision:"2e5e8f5eba25e2675b56d4a0ff12b70e"},{url:"/AvenirNextRoundedRegular-1080183-export/AvenirNextRoundedRegular-1080183.woff",revision:"b77d744fb2834b268a252ba377cdc6de"},{url:"/AvenirNextRoundedRegular-1080183-export/AvenirNextRoundedRegular-1080183.woff2",revision:"7e4a9b8881609db0f6a971bcc5a3ef39"},{url:"/_next/app-build-manifest.json",revision:"8c5329920217e2dece4012969d3f63cc"},{url:"/_next/static/ALvw0C-24BtSMY_n4Oish/_buildManifest.js",revision:"4e7cba7e679d15c6635270870fb7ba38"},{url:"/_next/static/ALvw0C-24BtSMY_n4Oish/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/1004.e8196906df9d2d37.js",revision:"e8196906df9d2d37"},{url:"/_next/static/chunks/1004.e8196906df9d2d37.js.map",revision:"22875d43a9eac4e0b326092b701ed1fd"},{url:"/_next/static/chunks/1313.d35a3ff320d0590c.js",revision:"d35a3ff320d0590c"},{url:"/_next/static/chunks/1313.d35a3ff320d0590c.js.map",revision:"e16d297a6012599f412472667597d855"},{url:"/_next/static/chunks/1379.237c4b79d8d7b465.js",revision:"237c4b79d8d7b465"},{url:"/_next/static/chunks/1379.237c4b79d8d7b465.js.map",revision:"dd3f687de2240e7a27599e75faa962be"},{url:"/_next/static/chunks/1431.3c48a3c16aab9f28.js",revision:"3c48a3c16aab9f28"},{url:"/_next/static/chunks/1431.3c48a3c16aab9f28.js.map",revision:"d3523ab8238be30098c126075a361af5"},{url:"/_next/static/chunks/1612.8cb7f823c221fd24.js",revision:"8cb7f823c221fd24"},{url:"/_next/static/chunks/1612.8cb7f823c221fd24.js.map",revision:"144fea08abfc3ad454d1d092a13edc31"},{url:"/_next/static/chunks/183.40f993f72fbe8d7e.js",revision:"40f993f72fbe8d7e"},{url:"/_next/static/chunks/183.40f993f72fbe8d7e.js.map",revision:"47e48c5d9040fafa4b85f65edc19bfa7"},{url:"/_next/static/chunks/1843.1225c1e9862eabd6.js",revision:"1225c1e9862eabd6"},{url:"/_next/static/chunks/1843.1225c1e9862eabd6.js.map",revision:"aa95849264dc3b7f97e225fbba20eef3"},{url:"/_next/static/chunks/18b16e15-c7fa9a10f549bc53.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/18b16e15-c7fa9a10f549bc53.js.map",revision:"4edb8068b4acecfa5edd70cb6f06a2bf"},{url:"/_next/static/chunks/2058.443982227ed90ff2.js",revision:"443982227ed90ff2"},{url:"/_next/static/chunks/2058.443982227ed90ff2.js.map",revision:"406bc63add6341c45ffee4e099653a75"},{url:"/_next/static/chunks/2084.721969af66f6a1b4.js",revision:"721969af66f6a1b4"},{url:"/_next/static/chunks/2084.721969af66f6a1b4.js.map",revision:"7074c54c1683cafca1f43c7b173dca7b"},{url:"/_next/static/chunks/2513.f4261cbc4c5ca2e7.js",revision:"f4261cbc4c5ca2e7"},{url:"/_next/static/chunks/2513.f4261cbc4c5ca2e7.js.map",revision:"a68fad71facfd946322c9f3118294463"},{url:"/_next/static/chunks/2527.3d63cba5a47c5542.js",revision:"3d63cba5a47c5542"},{url:"/_next/static/chunks/2527.3d63cba5a47c5542.js.map",revision:"862e8457f9bd000c7b9d7ea60ab74d42"},{url:"/_next/static/chunks/3009.28d3a81601309cf8.js",revision:"28d3a81601309cf8"},{url:"/_next/static/chunks/3009.28d3a81601309cf8.js.map",revision:"f2a7cfcf2eb2ad9c3a4a8ebd0f25ae5d"},{url:"/_next/static/chunks/3040-b8af43284a5e2655.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/3040-b8af43284a5e2655.js.map",revision:"c007570198379f21d61c79f6355d9378"},{url:"/_next/static/chunks/3106.3428e2aef5aae2fe.js",revision:"3428e2aef5aae2fe"},{url:"/_next/static/chunks/3106.3428e2aef5aae2fe.js.map",revision:"07152bd06cf2f705f4cb4ee677f3bc1a"},{url:"/_next/static/chunks/3178.14ed89786c7fdaf1.js",revision:"14ed89786c7fdaf1"},{url:"/_next/static/chunks/3178.14ed89786c7fdaf1.js.map",revision:"107a36cee548db39c7887e490e4adb13"},{url:"/_next/static/chunks/3470.738ba65f2e46a7e5.js",revision:"738ba65f2e46a7e5"},{url:"/_next/static/chunks/3470.738ba65f2e46a7e5.js.map",revision:"53877601621d7dc5e0e4e30313b8e118"},{url:"/_next/static/chunks/3485.fec580dec970a563.js",revision:"fec580dec970a563"},{url:"/_next/static/chunks/3485.fec580dec970a563.js.map",revision:"4c7ef04bdcc69881633dd45e3bec91da"},{url:"/_next/static/chunks/35.cc63778b09482458.js",revision:"cc63778b09482458"},{url:"/_next/static/chunks/35.cc63778b09482458.js.map",revision:"cf2846581eee5a591d8a4a592532dd9a"},{url:"/_next/static/chunks/3527-92c4af87d7544430.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/3527-92c4af87d7544430.js.map",revision:"ba4cf6ad5a625b3e247bb22af89f75af"},{url:"/_next/static/chunks/3761.f390d71a8c3a3f30.js",revision:"f390d71a8c3a3f30"},{url:"/_next/static/chunks/3761.f390d71a8c3a3f30.js.map",revision:"93b37ac1053354802ff83550e1decb96"},{url:"/_next/static/chunks/3793-42ecf4a38d9c88e8.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/3813.265bbdba1e5e3e50.js",revision:"265bbdba1e5e3e50"},{url:"/_next/static/chunks/3813.265bbdba1e5e3e50.js.map",revision:"e9e900d3d1f52e0895ff10318d7778cd"},{url:"/_next/static/chunks/3840.3a74d7a9662cf0f2.js",revision:"3a74d7a9662cf0f2"},{url:"/_next/static/chunks/3840.3a74d7a9662cf0f2.js.map",revision:"4222e68e33f85e8ce65597c156f23ceb"},{url:"/_next/static/chunks/3930.84d81ab69f8e184e.js",revision:"84d81ab69f8e184e"},{url:"/_next/static/chunks/3930.84d81ab69f8e184e.js.map",revision:"d41abe9eda22a9ebd03d9dd0fe5f857c"},{url:"/_next/static/chunks/4129.e10e5225439a8302.js",revision:"e10e5225439a8302"},{url:"/_next/static/chunks/4129.e10e5225439a8302.js.map",revision:"5eaca6406bf71103b870a9e6d4c614b2"},{url:"/_next/static/chunks/4311.9928c50563c60a8d.js",revision:"9928c50563c60a8d"},{url:"/_next/static/chunks/4311.9928c50563c60a8d.js.map",revision:"007cbb36ac60c6074c8c8aef20e084f5"},{url:"/_next/static/chunks/432.307faf39b8d17ac8.js",revision:"307faf39b8d17ac8"},{url:"/_next/static/chunks/432.307faf39b8d17ac8.js.map",revision:"32f15981a534ba4f3adb79b29d146cfe"},{url:"/_next/static/chunks/4390.456f64929fd37c44.js",revision:"456f64929fd37c44"},{url:"/_next/static/chunks/4390.456f64929fd37c44.js.map",revision:"5cb5eb19863ef582838c20cca419bfb9"},{url:"/_next/static/chunks/474.cf75ea0ef0d83bdc.js",revision:"cf75ea0ef0d83bdc"},{url:"/_next/static/chunks/474.cf75ea0ef0d83bdc.js.map",revision:"f450a6c433a394ac6d266353ef205f45"},{url:"/_next/static/chunks/4928.e06783c962e02092.js",revision:"e06783c962e02092"},{url:"/_next/static/chunks/4928.e06783c962e02092.js.map",revision:"997500d984bb206015942768dddb0af7"},{url:"/_next/static/chunks/5061.94bb1981ba994dda.js",revision:"94bb1981ba994dda"},{url:"/_next/static/chunks/5061.94bb1981ba994dda.js.map",revision:"e2f0019a3bdb7f91431333c8087ac116"},{url:"/_next/static/chunks/5154.0acb602e397df24b.js",revision:"0acb602e397df24b"},{url:"/_next/static/chunks/5154.0acb602e397df24b.js.map",revision:"f7fa2e58c5648b79ce6e33b0434afe55"},{url:"/_next/static/chunks/5262.ab52f5bb275e3203.js",revision:"ab52f5bb275e3203"},{url:"/_next/static/chunks/5262.ab52f5bb275e3203.js.map",revision:"57a654cb3bb40ddf4f97f4f59dd68e32"},{url:"/_next/static/chunks/5550-b4a926c97e2c5376.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/5550-b4a926c97e2c5376.js.map",revision:"00d4567a2830ab61c780afda5bc6d1e2"},{url:"/_next/static/chunks/5698.d407a957bdda6088.js",revision:"d407a957bdda6088"},{url:"/_next/static/chunks/5698.d407a957bdda6088.js.map",revision:"ab9b7d64851a2511a2c2b0c28ec17846"},{url:"/_next/static/chunks/5761.045122dd9e68eeda.js",revision:"045122dd9e68eeda"},{url:"/_next/static/chunks/5761.045122dd9e68eeda.js.map",revision:"ef1b6fdf13de7ee2a16c24099cd666a4"},{url:"/_next/static/chunks/5842.646ce968e2d41fa4.js",revision:"646ce968e2d41fa4"},{url:"/_next/static/chunks/5988.b0efe90b26384a5c.js",revision:"b0efe90b26384a5c"},{url:"/_next/static/chunks/5988.b0efe90b26384a5c.js.map",revision:"acf68c9785b9e0717f81c3def1c71b15"},{url:"/_next/static/chunks/6066.c31de60e4fe6f81f.js",revision:"c31de60e4fe6f81f"},{url:"/_next/static/chunks/6066.c31de60e4fe6f81f.js.map",revision:"7fc69b8d07d73b0e32df3b59bd277aee"},{url:"/_next/static/chunks/6248-819479c748b0970c.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/6248-819479c748b0970c.js.map",revision:"2f8b4cfa2ce9c14958eac607923e5fec"},{url:"/_next/static/chunks/6322-1090b8fd663b7c1a.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/6322-1090b8fd663b7c1a.js.map",revision:"e7e24567d1fc10d66b5eae8999106acc"},{url:"/_next/static/chunks/645.db654d3ed69e2e63.js",revision:"db654d3ed69e2e63"},{url:"/_next/static/chunks/645.db654d3ed69e2e63.js.map",revision:"185c719a039ef2c5c4712c04a034373d"},{url:"/_next/static/chunks/6457-0e3b931462b2dc1c.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/6457-0e3b931462b2dc1c.js.map",revision:"c56d68d0092b1bacd225956ef0ee6360"},{url:"/_next/static/chunks/6482.5ab67833f8c83b1d.js",revision:"5ab67833f8c83b1d"},{url:"/_next/static/chunks/6482.5ab67833f8c83b1d.js.map",revision:"e4824573c76dc09ac8058a124f7205f7"},{url:"/_next/static/chunks/6527.c8e348452104a7cf.js",revision:"c8e348452104a7cf"},{url:"/_next/static/chunks/6527.c8e348452104a7cf.js.map",revision:"78f01862509a6e11f11d0e7df2002c38"},{url:"/_next/static/chunks/6892.4411a0bc75f0cfe1.js",revision:"4411a0bc75f0cfe1"},{url:"/_next/static/chunks/6892.4411a0bc75f0cfe1.js.map",revision:"02e3334c915b52ec54621ea57baf2847"},{url:"/_next/static/chunks/6923.3d83b76f2a8d0253.js",revision:"3d83b76f2a8d0253"},{url:"/_next/static/chunks/6923.3d83b76f2a8d0253.js.map",revision:"446fa804e43d45f2f5257c8f475ab074"},{url:"/_next/static/chunks/7074.35929822a1976bef.js",revision:"35929822a1976bef"},{url:"/_next/static/chunks/7074.35929822a1976bef.js.map",revision:"22b38433242af691d811751c0cda1751"},{url:"/_next/static/chunks/7569.b7bd65468d3829ed.js",revision:"b7bd65468d3829ed"},{url:"/_next/static/chunks/7569.b7bd65468d3829ed.js.map",revision:"a698e8ed75d21018a03b868ce1da7803"},{url:"/_next/static/chunks/7583.7797feaa833f5d23.js",revision:"7797feaa833f5d23"},{url:"/_next/static/chunks/7583.7797feaa833f5d23.js.map",revision:"9d90b8b82140f5bfc4f5b247233a9c74"},{url:"/_next/static/chunks/7617.b8d2feff1c6d13a6.js",revision:"b8d2feff1c6d13a6"},{url:"/_next/static/chunks/7617.b8d2feff1c6d13a6.js.map",revision:"6bae3b355b34f0aa0354e6b64588ebf4"},{url:"/_next/static/chunks/7638-bc3914a887e17c66.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/7638-bc3914a887e17c66.js.map",revision:"1bb14bf53a3a3fca85c258df814227b1"},{url:"/_next/static/chunks/7788.3ad7ad9f1e7f7833.js",revision:"3ad7ad9f1e7f7833"},{url:"/_next/static/chunks/7788.3ad7ad9f1e7f7833.js.map",revision:"cf6c31d0b4727880dbb21e36c2b2cb5b"},{url:"/_next/static/chunks/7874.89fb9eeb6be31224.js",revision:"89fb9eeb6be31224"},{url:"/_next/static/chunks/7874.89fb9eeb6be31224.js.map",revision:"20f50cd0532442d9984ccf1a230bd356"},{url:"/_next/static/chunks/790.13f6c4b9ff1791fd.js",revision:"13f6c4b9ff1791fd"},{url:"/_next/static/chunks/790.13f6c4b9ff1791fd.js.map",revision:"43b0384d0af5578b72bf2353ce8248e5"},{url:"/_next/static/chunks/7945.8067e2fa1c882f97.js",revision:"8067e2fa1c882f97"},{url:"/_next/static/chunks/7945.8067e2fa1c882f97.js.map",revision:"25e7f04d95da029e45237921184045eb"},{url:"/_next/static/chunks/7e28f48e-5d0313d8310c9156.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/7e28f48e-5d0313d8310c9156.js.map",revision:"d144d0618337dd9843b905d705260304"},{url:"/_next/static/chunks/8057.aa273303cb883692.js",revision:"aa273303cb883692"},{url:"/_next/static/chunks/8057.aa273303cb883692.js.map",revision:"a0e76f12b7f9009a66ea3e8d6f56b584"},{url:"/_next/static/chunks/8250.83b4198009f06312.js",revision:"83b4198009f06312"},{url:"/_next/static/chunks/8250.83b4198009f06312.js.map",revision:"afb6d5be266b24f2f078b246bdee6a93"},{url:"/_next/static/chunks/830.bae1ca90def593f2.js",revision:"bae1ca90def593f2"},{url:"/_next/static/chunks/830.bae1ca90def593f2.js.map",revision:"105063600626d28f39239bd788a10f4d"},{url:"/_next/static/chunks/8343-2e3cf6067985cdf1.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/8343-2e3cf6067985cdf1.js.map",revision:"9351517238251e3ff86c5d4fbc44945b"},{url:"/_next/static/chunks/8456.25e5d2cb26ee031a.js",revision:"25e5d2cb26ee031a"},{url:"/_next/static/chunks/8456.25e5d2cb26ee031a.js.map",revision:"3296afb797c8d246ab3c7291ce345a67"},{url:"/_next/static/chunks/8517.5d60ccbb8c18c647.js",revision:"5d60ccbb8c18c647"},{url:"/_next/static/chunks/8517.5d60ccbb8c18c647.js.map",revision:"8096cfaed153334962c681831d0ea56a"},{url:"/_next/static/chunks/8624.69bca3b30ceeab59.js",revision:"69bca3b30ceeab59"},{url:"/_next/static/chunks/8624.69bca3b30ceeab59.js.map",revision:"8710a1c28bd36689244545e69e87ecd4"},{url:"/_next/static/chunks/8750.be187129cb2ee251.js",revision:"be187129cb2ee251"},{url:"/_next/static/chunks/8750.be187129cb2ee251.js.map",revision:"90f65370dba730445d90afcd04265d60"},{url:"/_next/static/chunks/8791.1fa8b4b35170ab7c.js",revision:"1fa8b4b35170ab7c"},{url:"/_next/static/chunks/8791.1fa8b4b35170ab7c.js.map",revision:"f378c498f166e67e104e9ad05eb373fe"},{url:"/_next/static/chunks/8792-bc34535a50167d19.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/8792-bc34535a50167d19.js.map",revision:"7c4bf70cfcd005fe55232de38f71f6a1"},{url:"/_next/static/chunks/8997.9487f2b0653c146f.js",revision:"9487f2b0653c146f"},{url:"/_next/static/chunks/8997.9487f2b0653c146f.js.map",revision:"9d71c4c256f17077d12824c072afd9ae"},{url:"/_next/static/chunks/9058.681a93fdcfdf8ef5.js",revision:"681a93fdcfdf8ef5"},{url:"/_next/static/chunks/9058.681a93fdcfdf8ef5.js.map",revision:"4d89158fa99909704b5c68a8af7ba690"},{url:"/_next/static/chunks/9154-790f23abccdb3e49.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/9154-790f23abccdb3e49.js.map",revision:"8e2d49c5257d72b24ec4f6ed3e9e93fb"},{url:"/_next/static/chunks/9454-a41fa53771eb5a14.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/9454-a41fa53771eb5a14.js.map",revision:"05f327df0fe688a3a69f1ec5607b5be4"},{url:"/_next/static/chunks/9536.220bd442696c7036.js",revision:"220bd442696c7036"},{url:"/_next/static/chunks/9536.220bd442696c7036.js.map",revision:"0dfefbae126fa34db207d499ad650179"},{url:"/_next/static/chunks/9631.d471117bd53d2717.js",revision:"d471117bd53d2717"},{url:"/_next/static/chunks/9631.d471117bd53d2717.js.map",revision:"6fbfe58113aa0685f6c667a3e34ee3b5"},{url:"/_next/static/chunks/9678-778ff3b23c04f64f.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/9678-778ff3b23c04f64f.js.map",revision:"7ae7be6082a2c3dbb7da8991d4c5aa1e"},{url:"/_next/static/chunks/9989.a8b158d927b15331.js",revision:"a8b158d927b15331"},{url:"/_next/static/chunks/9989.a8b158d927b15331.js.map",revision:"86a7b22cdf1a6922310a7a739e986cfc"},{url:"/_next/static/chunks/a31ab248-eed7c6516d748d54.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/a31ab248-eed7c6516d748d54.js.map",revision:"af2c5539be46e9b7ce28f3efaa83bfd1"},{url:"/_next/static/chunks/app/_not-found/page-f27100c7951df8d1.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/app/bots/page-a33f775e6f56fd98.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/app/bots/page-a33f775e6f56fd98.js.map",revision:"87f313bedba7420bedce78bbd87b7012"},{url:"/_next/static/chunks/app/error-1ebdad38a45beb59.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/app/error-1ebdad38a45beb59.js.map",revision:"85276708ce8e8b27ee9d264c6905b9ca"},{url:"/_next/static/chunks/app/forecasting/%5BchainId%5D/%5BmarketAddress%5D/page-1f6ab2d71e283606.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/app/forecasting/%5BchainId%5D/%5BmarketAddress%5D/page-1f6ab2d71e283606.js.map",revision:"5e4a4471e08c9c5fd4823b66f78c3c4b"},{url:"/_next/static/chunks/app/forecasting/page-3768cd7985c5f86b.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/app/forecasting/page-3768cd7985c5f86b.js.map",revision:"f7abd32f6b204e5ef4c8ba2a8d113489"},{url:"/_next/static/chunks/app/futarchy/page-fdb03b2f6a63e741.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/app/futarchy/page-fdb03b2f6a63e741.js.map",revision:"408995ffd11a8321151591c30f9dea9c"},{url:"/_next/static/chunks/app/global-error-918feb010657bae6.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/app/global-error-918feb010657bae6.js.map",revision:"c93f71e90d79a6aa53384d08ac9d9bb7"},{url:"/_next/static/chunks/app/layout-5f1adbca6810dbe0.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/app/layout-5f1adbca6810dbe0.js.map",revision:"b2d805618948e31581750e19908eb839"},{url:"/_next/static/chunks/app/leaderboard/%5Bid%5D/epochs/%5Bepoch%5D/page-d58ba1aadc6a6bab.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/app/leaderboard/%5Bid%5D/epochs/%5Bepoch%5D/page-d58ba1aadc6a6bab.js.map",revision:"9fd49c748bb13a6f09fdc97a830aee27"},{url:"/_next/static/chunks/app/leaderboard/page-07e7bfb4f8a0755f.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/app/leaderboard/page-07e7bfb4f8a0755f.js.map",revision:"281f5b02707f63373d3215a95fecd19d"},{url:"/_next/static/chunks/app/not-found-d03852b57c10b5ea.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/app/page-7301ddd33b4ca493.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/app/page-7301ddd33b4ca493.js.map",revision:"c88772dbecc11f4f30297171f49bef2e"},{url:"/_next/static/chunks/f4dc6ce0-7d93a3b742b388d5.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/f4dc6ce0-7d93a3b742b388d5.js.map",revision:"b41ec79bb0c7bec23559ecbd03b2c9a7"},{url:"/_next/static/chunks/framework-1625a25bd56a487b.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/framework-1625a25bd56a487b.js.map",revision:"b61dec9ba5e7fc11d2aac637da9a1856"},{url:"/_next/static/chunks/main-app-8a1fbcc3ac87195b.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/main-app-8a1fbcc3ac87195b.js.map",revision:"87a99e157b6e4d02ecb6300af22ec77b"},{url:"/_next/static/chunks/main-b66bab6e4ca24744.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/main-b66bab6e4ca24744.js.map",revision:"2baed322d168e42b8b8ebafa1bdf3820"},{url:"/_next/static/chunks/pages/_app-1578276777346936.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/pages/_app-1578276777346936.js.map",revision:"a737047ee188560b9ea1121d4d883f54"},{url:"/_next/static/chunks/pages/_error-456bc4d02b78646c.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/pages/_error-456bc4d02b78646c.js.map",revision:"ad96e20c79d82bfb66cc1842e813cde9"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-36a654dccdbc4df6.js",revision:"ALvw0C-24BtSMY_n4Oish"},{url:"/_next/static/chunks/webpack-36a654dccdbc4df6.js.map",revision:"a62d43b6d9a60519338d6e8c371aa9f3"},{url:"/_next/static/css/01e303d912fdd7db.css",revision:"01e303d912fdd7db"},{url:"/_next/static/css/01e303d912fdd7db.css.map",revision:"e2968a09346b76b0991e5b24bfe65214"},{url:"/_next/static/css/8073d4869bb3d095.css",revision:"8073d4869bb3d095"},{url:"/_next/static/css/8073d4869bb3d095.css.map",revision:"f7936320d2959b9143c53bf2cd67acad"},{url:"/_next/static/css/8268cb36a23698b6.css",revision:"8268cb36a23698b6"},{url:"/_next/static/css/8268cb36a23698b6.css.map",revision:"0ff111f3094ec74b08324d5ecf872a4f"},{url:"/_next/static/media/font.0d99608a.woff",revision:"0d99608a"},{url:"/_next/static/media/font.1437a347.woff2",revision:"1437a347"},{url:"/_next/static/media/font.15f0cb14.woff2",revision:"15f0cb14"},{url:"/_next/static/media/font.172826d2.woff2",revision:"172826d2"},{url:"/_next/static/media/font.21d6c94e.woff2",revision:"21d6c94e"},{url:"/_next/static/media/font.2484265e.woff",revision:"2484265e"},{url:"/_next/static/media/font.248fd520.woff",revision:"248fd520"},{url:"/_next/static/media/font.2913dd65.woff",revision:"2913dd65"},{url:"/_next/static/media/font.29904e4a.woff2",revision:"29904e4a"},{url:"/_next/static/media/font.2c3dda54.woff",revision:"2c3dda54"},{url:"/_next/static/media/font.2d75bfb1.woff2",revision:"2d75bfb1"},{url:"/_next/static/media/font.2f474a58.woff2",revision:"2f474a58"},{url:"/_next/static/media/font.2f548f0d.woff2",revision:"2f548f0d"},{url:"/_next/static/media/font.2fa17060.woff2",revision:"2fa17060"},{url:"/_next/static/media/font.3057c2b7.woff",revision:"3057c2b7"},{url:"/_next/static/media/font.3108bbad.woff",revision:"3108bbad"},{url:"/_next/static/media/font.31ed5aac.woff2",revision:"31ed5aac"},{url:"/_next/static/media/font.37cdad6c.woff2",revision:"37cdad6c"},{url:"/_next/static/media/font.39d2f5c6.woff2",revision:"39d2f5c6"},{url:"/_next/static/media/font.435bd86c.woff2",revision:"435bd86c"},{url:"/_next/static/media/font.4b393fab.woff",revision:"4b393fab"},{url:"/_next/static/media/font.50bbef7d.woff",revision:"50bbef7d"},{url:"/_next/static/media/font.553b8400.woff",revision:"553b8400"},{url:"/_next/static/media/font.5cf8a729.woff2",revision:"5cf8a729"},{url:"/_next/static/media/font.5d6a72af.woff2",revision:"5d6a72af"},{url:"/_next/static/media/font.635bb62c.woff",revision:"635bb62c"},{url:"/_next/static/media/font.6509c247.woff",revision:"6509c247"},{url:"/_next/static/media/font.6cd92d24.woff2",revision:"6cd92d24"},{url:"/_next/static/media/font.73a9d55f.woff",revision:"73a9d55f"},{url:"/_next/static/media/font.7bcc2af9.woff",revision:"7bcc2af9"},{url:"/_next/static/media/font.7c4c2bf8.woff2",revision:"7c4c2bf8"},{url:"/_next/static/media/font.7c7cecdb.woff2",revision:"7c7cecdb"},{url:"/_next/static/media/font.816c5319.woff",revision:"816c5319"},{url:"/_next/static/media/font.82bbad3b.woff2",revision:"82bbad3b"},{url:"/_next/static/media/font.856572dd.woff2",revision:"856572dd"},{url:"/_next/static/media/font.8bb377a0.woff",revision:"8bb377a0"},{url:"/_next/static/media/font.8ee62147.woff2",revision:"8ee62147"},{url:"/_next/static/media/font.8f4ab0ae.woff2",revision:"8f4ab0ae"},{url:"/_next/static/media/font.9491bc5f.woff2",revision:"9491bc5f"},{url:"/_next/static/media/font.9c12eced.woff",revision:"9c12eced"},{url:"/_next/static/media/font.9d17a49d.woff",revision:"9d17a49d"},{url:"/_next/static/media/font.9fb5eaf7.woff",revision:"9fb5eaf7"},{url:"/_next/static/media/font.9fd6a2c2.woff",revision:"9fd6a2c2"},{url:"/_next/static/media/font.a11bbcef.woff",revision:"a11bbcef"},{url:"/_next/static/media/font.a3e54e3d.woff",revision:"a3e54e3d"},{url:"/_next/static/media/font.a4380fcb.woff",revision:"a4380fcb"},{url:"/_next/static/media/font.a7183547.woff2",revision:"a7183547"},{url:"/_next/static/media/font.a7d755eb.woff",revision:"a7d755eb"},{url:"/_next/static/media/font.afa2688e.woff",revision:"afa2688e"},{url:"/_next/static/media/font.b3f27314.woff",revision:"b3f27314"},{url:"/_next/static/media/font.b850aef4.woff2",revision:"b850aef4"},{url:"/_next/static/media/font.b869ae38.woff",revision:"b869ae38"},{url:"/_next/static/media/font.b99947cf.woff",revision:"b99947cf"},{url:"/_next/static/media/font.ba404641.woff2",revision:"ba404641"},{url:"/_next/static/media/font.bdffa644.woff2",revision:"bdffa644"},{url:"/_next/static/media/font.be02cff3.woff",revision:"be02cff3"},{url:"/_next/static/media/font.bf00c3fd.woff2",revision:"bf00c3fd"},{url:"/_next/static/media/font.c0a46350.woff2",revision:"c0a46350"},{url:"/_next/static/media/font.c9641110.woff",revision:"c9641110"},{url:"/_next/static/media/font.cb0db050.woff2",revision:"cb0db050"},{url:"/_next/static/media/font.d0a69e17.woff2",revision:"d0a69e17"},{url:"/_next/static/media/font.d18ef773.woff",revision:"d18ef773"},{url:"/_next/static/media/font.d3118412.woff2",revision:"d3118412"},{url:"/_next/static/media/font.d63f196b.woff2",revision:"d63f196b"},{url:"/_next/static/media/font.d8425c1f.woff2",revision:"d8425c1f"},{url:"/_next/static/media/font.d890d2d6.woff",revision:"d890d2d6"},{url:"/_next/static/media/font.e2b7d353.woff2",revision:"e2b7d353"},{url:"/_next/static/media/font.e6399456.woff2",revision:"e6399456"},{url:"/_next/static/media/font.e7691173.woff2",revision:"e7691173"},{url:"/_next/static/media/font.e93302c2.woff",revision:"e93302c2"},{url:"/_next/static/media/font.e985e56f.woff",revision:"e985e56f"},{url:"/_next/static/media/font.e9e26597.woff2",revision:"e9e26597"},{url:"/_next/static/media/font.f0702b0f.woff",revision:"f0702b0f"},{url:"/_next/static/media/font.f1b46486.woff2",revision:"f1b46486"},{url:"/_next/static/media/font.f2fca399.woff",revision:"f2fca399"},{url:"/_next/static/media/font.f6457293.woff",revision:"f6457293"},{url:"/_next/static/media/font.f8d712aa.woff",revision:"f8d712aa"},{url:"/_next/static/media/font.faad02c9.woff",revision:"faad02c9"},{url:"/_next/static/media/font.fd1ecef3.woff",revision:"fd1ecef3"},{url:"/_next/static/media/font.fffd047e.woff2",revision:"fffd047e"},{url:"/claude.mp4",revision:"c8753825822cb0d863e9cb356b828217"},{url:"/etherscan.svg",revision:"7b7de5f72960e6da7ea41f35c3bd19da"},{url:"/favicon.svg",revision:"6f9c4e4689886cdc55e2ba1c52ce6cad"},{url:"/footer.png",revision:"7e3d75771110e9747f3a36dcbe4e9d2c"},{url:"/futarchy.png",revision:"d9695200cbd9d9d3eb25fbb8968b5b27"},{url:"/icons/icon-128x128.png",revision:"47b15e1ccdc4cab159cd43bfd1b8a208"},{url:"/icons/icon-144x144.png",revision:"91bd0b770d8c08e0517108dd20b5e09a"},{url:"/icons/icon-152x152.png",revision:"92e01007277fe4078fc585ad021f121f"},{url:"/icons/icon-192x192.png",revision:"007db62f745fd4b1cb63a1bac81d7b66"},{url:"/icons/icon-384x384.png",revision:"e481950d9351fa3b5149db414dff4144"},{url:"/icons/icon-512x512.png",revision:"97e635e3cc9a470518cd09ae80b2045d"},{url:"/icons/icon-72x72.png",revision:"3917754c4d089732711aa1324af449ff"},{url:"/icons/icon-96x96.png",revision:"44c96882dfc1f28a48b7fab96ebde9a6"},{url:"/logo.svg",revision:"fa07e01a214f11b798a83d5e5d79edc3"},{url:"/logomark.svg",revision:"bd5157f0874637ad15dae13d8893c8e4"},{url:"/lottie/loader.json",revision:"88d275ee06f4d3a8305299f8dcefc5fa"},{url:"/lottie/logomark.json",revision:"daab6abdc2aae8e70dd5aee7491672c3"},{url:"/manifest.json",revision:"a7d93b6cb0ee35bcc458602e953fda9f"},{url:"/og-image.png",revision:"a35a45c8a2a35e3df1fb612ea7a47e50"},{url:"/robots.txt",revision:"375f55efde6139766b4966f3299897d0"},{url:"/sitemap-0.xml",revision:"e4ef3c1ab973c940725d938b2b8ea3b9"},{url:"/sitemap.xml",revision:"ebbca34a7cd10f5467aecf3331106bcf"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:a,event:i,state:c})=>a&&"opaqueredirect"===a.type?new Response(a.body,{status:200,statusText:"OK",headers:a.headers}):a}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const a=e.pathname;return!a.startsWith("/api/auth/")&&!!a.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
 //# sourceMappingURL=sw.js.map
