+++ packages/sapience/src/components/FutarchyClientContent.tsx
@@ -1,0 +1,63 @@
+'use client';
+
+import { useEffect, useRef } from 'react';
+
+export default function FutarchyClientContent() {
+  const iframeRef = useRef<HTMLIFrameElement>(null);
+
+  // Force light mode rendering for the iframe
+  useEffect(() => {
+    const handleIframeLoad = () => {
+      const iframe = iframeRef.current;
+      // Guard already exists here, but keeping it doesn't hurt
+      if (typeof document === 'undefined') return;
+      if (iframe && iframe.contentDocument) {
+        try {
+          // Try to inject a style element to force light mode
+          const style = iframe.contentDocument.createElement('style');
+          style.textContent =
+            'html { color-scheme: light !important; } * { filter: none !important; }';
+          iframe.contentDocument.head.appendChild(style);
+        } catch (e) {
+          // Security policy might prevent this
+          console.error('Could not inject styles into iframe:', e);
+        }
+      }
+    };
+
+    const iframe = iframeRef.current;
+    if (iframe) {
+      // Ensure load event listener is attached only once iframe exists
+      iframe.addEventListener('load', handleIframeLoad);
+      // Clean up listener on unmount
+      return () => iframe.removeEventListener('load', handleIframeLoad);
+    }
+  }, []); // Empty dependency array ensures this runs once client-side
+
+  return (
+    <div
+      className="fixed inset-0 z-50 light w-[100dwv] pointer-events-none"
+      style={{
+        colorScheme: 'light',
+        filter: 'none',
+      }}
+    >
+      {/* eslint-disable-next-line jsx-a11y/iframe-has-title */}
+      <iframe
+        ref={iframeRef}
+        src="https://my.spline.design/particlesfutarchy-SDhuN0OYiCRHRPt2fFec4bCm/"
+        style={{
+          opacity: 0.5,
+          width: '100%',
+          height: '100%',
+          border: 'none',
+          colorScheme: 'light',
+          filter: 'none',
+        }}
+        loading="lazy"
+        referrerPolicy="no-referrer"
+        sandbox="allow-same-origin allow-scripts allow-downloads allow-forms allow-modals allow-orientation-lock allow-pointer-lock allow-popups allow-popups-to-escape-sandbox allow-presentation allow-storage-access-by-user-activation allow-top-navigation-by-user-activation"
+      />
+    </div>
+  );
+}
